refactor embedding button component
